added strlcat and updated int types to size_t in a few functions

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: no-conne <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 09:37:08 by no-conne          #+#    #+#             */
-/*   Updated: 2019/05/21 10:04:33 by no-conne         ###   ########.fr       */
+/*   Updated: 2019/05/21 10:29:17 by no-conne         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 char	*ft_strncat(char *dest, const char *src, size_t n)
 {
-	int i;
+	size_t i;
 	size_t x;
 
 	i = ft_strlen(dest);
